# Optimize AsyncClientManager by moving hashing to control plane