--- a/source/common/grpc/async_client_manager_impl.cc
+++ b/source/common/grpc/async_client_manager_impl.cc
@@ -49,11 +49,6 @@ AsyncClientFactoryImpl::AsyncClientFactoryImpl(Upstream::ClusterManager& cm,
   cm_.checkActiveStaticCluster(config.envoy_grpc().cluster_name());
 }
 
-GrpcServiceHashKeyWrapper::GrpcServiceHashKeyWrapper(
-    const envoy::config::core::v3::GrpcService& config)
-    : config_(config), pre_computed_hash_(Envoy::MessageUtil::hash(config)),
-      config_as_string_(Envoy::MessageUtil::getYamlStringFromMessage(config)) {}
-
 AsyncClientManagerImpl::AsyncClientManagerImpl(Upstream::ClusterManager& cm,
                                                ThreadLocal::Instance& tls, TimeSource& time_source,
                                                Api::Api& api, const StatNames& stat_names)