@DiazAlan 2023-08-23
@ggreenway This is a follow up from what we talked about on [28272](https://github.com/envoyproxy/envoy/pull/28272). This PR wraps the precomputed hash and config and sends that to the async client manager. I haven't changed the other filters yet as I want to get feedback on the initial implementation of this. If this looks good and passes, I'll make a follow up PR for the other filters.